Load benchmark results from disk
